+++ CHANGELOG.md
@@ -4,6 +4,6 @@
 - added docs page
 - reorganised methods
 - fix taiko rank calculations
-- mod calculations use string arrays instead of strings
+- mod calculations use string arrays instead of strings ("HDDTHR" -> ["HD", "DT", "HR"])
 
 ### 1.0.0